Modularize Sequelize and made it seperate from server
