+++ back-end/db/models/assignedBadges.js
@@ -1,4 +1,4 @@
-const sequelize = require("../../server");
+const sequelize = require("../sequelize");
 const { DataTypes } = require("sequelize");
 const { Badge } = require("./badges");
 const { User } = require("./users");
