+++ back-end/db/models/users.js
@@ -1,4 +1,4 @@
-const sequelize = require("../../server");
+const sequelize = require("../sequelize");
 const { DataTypes } = require("sequelize");
 const { Badge } = require("./badges");
 // const bcrypt = require("bcrypt");
