+++ back-end/db/sequelize.js
@@ -1,0 +1,14 @@
+const { Sequelize } = require("sequelize");
+require("dotenv").config();
+
+const sequelize = new Sequelize(
+  `${process.env.DB_NAME}`,
+  `${process.env.DB_USER}`,
+  `${process.env.DB_PASS}`,
+  {
+    host: `${process.env.DB_HOST}`,
+    dialect: "postgres",
+  }
+);
+
+module.exports = sequelize;
